refactor crate structure
