+++ crates/maimap-server/Cargo.toml
@@ -1,0 +1,31 @@
+[package]
+name = "maimap-server"
+version.workspace = true
+authors.workspace = true
+edition.workspace = true
+description.workspace = true
+repository.workspace = true
+readme.workspace = true
+license.workspace = true
+
+[[test]]
+name = "server-arcade-api-test"
+path = "tests/arcade_api.rs"
+
+[dependencies]
+maimap-utils = { workspace = true }
+salvo = { version = "0.78", features = ["test"] }
+tokio = { version = "1", features = ["macros", "test-util"] }
+tracing = "0.1"
+tracing-subscriber = "0.3"
+serde = { version = "1", features = ["derive"] }
+serde_json = "1"
+dotenv = "0.15.0"
+
+
+
+
+
+
+[target.'cfg(unix)'.dependencies]
+openssl = { version = "0.10", features = ["vendored"] }
