+++ crates/maimap-server/src/lib.rs
@@ -1,0 +1,3 @@
+pub mod handler;
+pub mod res;
+pub mod router;
