+++ crates/maimap-utils/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "maimap-utils"
+version.workspace = true
+authors.workspace = true
+edition.workspace = true
+description.workspace = true
+repository.workspace = true
+readme.workspace = true
+license.workspace = true
+
+[dependencies]
+serde = "1.0"
+serde_json = "1.0"
+dotenvy = "0.15"
+mongodb = "3.2"
+tracing = "0.1.41"
+ali-oss-rs = { version = "0.2" }
+thiserror = "2.0"
+anyhow = "1.0"
+
+[lints]
+workspace = true
