+++ crates/maimap-utils/src/errors.rs
@@ -1,3 +1,4 @@
+pub use anyhow::{Context, Error, Result};
 use mongodb::error::Error as MongoError;
 use std::io;
 use std::process::ExitStatus;
