--- src/lib.rs
@@ -1,16 +1,0 @@
-#[cfg(feature = "mongodb")]
-pub mod backup;
-#[cfg(feature = "mongodb")]
-pub mod db;
-pub mod env;
-
-#[cfg(feature = "mongodb")]
-pub mod errors;
-#[cfg(any(feature = "salvo"))]
-pub mod handler;
-pub mod res;
-
-#[cfg(any(feature = "salvo"))]
-pub mod router;
-#[cfg(feature = "mongodb")]
-pub mod types;
